// apps/web/src/shared/api/v1.gen.ts
export interface components {
    schemas: {
// ...
        Project: {
            /** Format: uuid */
            id: string;
            key: string;
            name: string;
            default_branch: string;
            latest_verdict: string;
            /** Format: int32 */
            total_findings: number;
            /** Format: date-time */
            created_at: string;
        };
        ProjectDetail: {
            /** Format: uuid */
            id: string;
            key: string;
            name: string;
            default_branch: string;
            latest_verdict: string;
            /** Format: int32 */
            total_findings: number;
            /** Format: date-time */
            created_at: string;
            target_pattern: string;
            languages: string[];
            quality_gate_rules: components["schemas"]["ProjectRuleView"][];
            severity_counts: {
                [key: string]: number;
            };
        };
        ProjectRuleView: {
            subtype: string;
            strategy_type: string;
        };
// ...
    };
// ...
}